adding typings
