+++ package.json
@@ -38,6 +38,7 @@
     "each": "^1.2.2",
     "mocha": "^7.1.0",
     "should": "^13.2.3",
+    "tslint": "^6.1.3",
     "typescript": "^4.4.2"
   },
   "bugs": {
